Full crab fight - external
Can die, noted odds in crab attacks, player has consistent damage. Can restart after death, can woo crab by talking to it. Continues with the rest of the program once crab dies. This is to be considered nearly complete with minor refinements to be added on per request of teammates (I will admit I am not the most creative/artsy)